需求[#32705]: SR-20210103-02-64 版本代码优化
    [描述说明]:  不写入日志优化
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/user/service/SysUserService.java
@@ -94,6 +94,12 @@ public class SysUserService {
         Long userIds = Long.valueOf(request.getHeader("userId"));
         String userName = String.valueOf(redisTemplate.opsForHash().get("userInfo:" + userIds, "userName"));
         String userNames = String.valueOf(redisTemplate.opsForHash().get("userInfo:" + userId, "userName"));
+        String userRole = String.valueOf(redisTemplate.opsForHash().entries("userInfo:"+userId).get("roleId"));
+        if(!"1234".equals(userRole)){
+            //权限不够；
+            throw new BusinessException(10008,"用户无权限");
+            //return -1;
+        }
         redisTemplate.delete("userInfo:" + userId);
         this.SysUserBackUpDao.deleteByPrimaryId(userId);
         MultiValueMap<String, Object> params = new LinkedMultiValueMap<>();
@@ -116,6 +122,12 @@ public class SysUserService {
     public int update(SysUser sysUser, HttpServletRequest request) throws IOException {
         Long userId = Long.valueOf(request.getHeader("userId"));
         String userName = String.valueOf(redisTemplate.opsForHash().get("userInfo:" + userId, "userName"));
+        String userRole = String.valueOf(redisTemplate.opsForHash().entries("userInfo:"+userId).get("roleId"));
+        if(!"1234".equals(userRole)){
+            //权限不够；
+            throw new BusinessException(10008,"用户无权限");
+            //return -1;
+        }
         SysUser sysUserCurrent = sysUserDao.selectByPrimaryId(sysUser.getUserId());
         StringBuilder sb = new StringBuilder();
         Long roleId = sysUser.getRoleId();
